# docs/swagger.yaml
basePath: /api
definitions:
...
  core.CreateChatGroupReq:
    properties:
      chat_name:
        type: string
    required:
    - chat_name
    type: object
  core.CreateDefaultChatReq:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
...
  core.SendMessageReq:
    properties:
      chat_id:
        type: integer
      text:
        type: string
    required:
    - chat_id
    - text
    type: object
  core.UpdateGroupChatAdminReq:
    properties:
      chat_id:
        type: integer
      new_admin_id:
        type: integer
    required:
    - chat_id
    - new_admin_id
    type: object
...
  core.WallChatResp:
    properties:
      chat_id:
        type: integer
      chat_last_message:
        type: string
      chat_name:
        type: string
    type: object
  rest.errorResponse:
    properties:
      error:
        type: string
    type: object
...
host: 35.180.225.96:8080
info:
  contact: {}
  description: API Server
  title: Social Network Backend
paths:
...
  /api/chat/default/create:
    post:
      consumes:
      - application/json
      description: create chat default
      operationId: createChatDefault
      parameters:
      - description: create chat default
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.CreateDefaultChatReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateChatDefault
      tags:
      - Chat
  /api/chat/default/delete/{chatId}:
    delete:
      description: delete chat default
      operationId: deleteChatDefault
      parameters:
      - description: chat id
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteChatDefault
      tags:
      - Chat
  /api/chat/group/admin/delete/{chatId}:
    delete:
      description: delete chat group
      operationId: deleteChatGroup
      parameters:
      - description: chat id
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteChatGroup
      tags:
      - Chat
  /api/chat/group/admin/update:
    put:
      consumes:
      - application/json
      description: transfer chat group admin
      operationId: transferChatGroupAdmin
      parameters:
      - description: update chat group admin
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.UpdateGroupChatAdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: TransferChatGroupAdmin
      tags:
      - Chat
  /api/chat/group/create:
    post:
      consumes:
      - application/json
      description: create chat group
      operationId: createChatGroup
      parameters:
      - description: create chat group
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.CreateChatGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateChatGroup
      tags:
      - Chat
  /api/chat/group/join/{chatId}:
    post:
      description: join chat
      operationId: joinChat
      parameters:
      - description: chat id
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: JoinChat
      tags:
      - Chat
  /api/chat/group/leave/{chatId}:
    delete:
      description: leave chat group
      operationId: leaveChatGroup
      parameters:
      - description: chat id
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: LeaveChatGroup
      tags:
      - Chat
  /api/chat/message/get/{chatId}:
    get:
      description: get messages from chat
      operationId: getMessages
      parameters:
      - description: chat id
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetMessages
      tags:
      - Chat
  /api/chat/message/send:
    post:
      consumes:
      - application/json
      description: send message
      operationId: sendMessage
      parameters:
      - description: message
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/core.SendMessageReq'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: SendMessage
      tags:
      - Chat
  /api/chat/wall:
    get:
      description: get chat wall
      operationId: chatWall
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.WallChatResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: ChatWall
      tags:
      - Chat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"